requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# PDF Processing
PyPDF2==3.0.1

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Data Processing
pydantic==2.5.0

# AWS DynamoDB 
boto3==1.29.7
